+++ apps/backend/src/user/user.controller.spec.ts
@@ -1,0 +1,30 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { UserController } from './user.controller';
+import { UserService } from './user.service';
+import { DatabaseModule } from '../database/database.module';
+import { JwtModule } from '@nestjs/jwt';
+
+describe('UserController', () => {
+  let controller: UserController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [UserController],
+      providers: [UserService],
+      imports: [
+        DatabaseModule,
+        JwtModule.register({
+          secret: process.env.JWT_KEY,
+          secretOrPrivateKey: process.env.JWT_KEY,
+          signOptions: { expiresIn: '1h' },
+        }),
+      ],
+    }).compile();
+
+    controller = module.get<UserController>(UserController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});