+++ apps/backend/src/user/user.service.spec.ts
@@ -1,0 +1,28 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { UserService } from './user.service';
+import { DatabaseModule } from '../database/database.module';
+import { JwtModule } from '@nestjs/jwt';
+
+describe('UsersService', () => {
+  let service: UserService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [UserService],
+      imports: [
+        DatabaseModule,
+        JwtModule.register({
+          secret: process.env.JWT_KEY,
+          secretOrPrivateKey: process.env.JWT_KEY,
+          signOptions: { expiresIn: '1h' },
+        }),
+      ],
+    }).compile();
+
+    service = module.get<UserService>(UserService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});